+++ src/components/GoogleAddressLookup/component.js
@@ -10,6 +10,9 @@ import getSuggestions from './helpers/getSuggestions';
 import getSearchForOption from './helpers/getSearchForOption';
 import getFormattedValue from './helpers/getFormattedValue';
 import SelectedLocationIcon from './icons/selectedLocationIcon';
+import RenderIf from '../RenderIf';
+import getSearchParams from './helpers/getSearchParams';
+import * as CustomPropTypes from './proptypes';
 import './styles.css';
 
 class PlacesLookupComponent extends Component {
@@ -21,7 +24,6 @@ class PlacesLookupComponent extends Component {
         this.handleSearch = this.handleSearch.bind(this);
 
         this.state = {
-            initialized: false,
             isSearching: false,
             places: [],
             suggestions: [],
@@ -64,10 +66,10 @@ class PlacesLookupComponent extends Component {
     handleSearch(value) {
         if (!this.initialized) return;
 
-        if (value && this.state.searchValue !== value) {
+        if (value) {
             this.searchGooglePlaces(value)
                 .then(results => this.processSearchResults(results, value))
-                .catch(this.handleError(value));
+                .catch(() => this.handleError(value));
         } else if (!value) {
             this.clear();
         }
@@ -75,39 +77,8 @@ class PlacesLookupComponent extends Component {
 
     searchGooglePlaces(needle) {
         const { searchOptions } = this.props;
-        let searchParams = {};
-        if (searchOptions) {
-            const { bounds, country, location, ...otherSearchOptions } = searchOptions;
+        const searchParams = getSearchParams(searchOptions);
 
-            if (bounds) {
-                const {
-                    sw: { latitude: swLatitude, longitude: swLongitude },
-                    ne: { latitude: neLatitude, longitude: neLongitude },
-                } = bounds;
-                searchParams = Object.assign(searchParams, {
-                    bounds: new window.google.maps.LatLngBounds(
-                        new window.google.maps.LatLng(swLatitude, swLongitude),
-                        new window.google.maps.LatLng(neLatitude, neLongitude),
-                    ),
-                });
-            }
-
-            if (location) {
-                searchParams = Object.assign(searchParams, {
-                    location: new window.google.maps.LatLng(location.latitude, location.longitude),
-                });
-            }
-
-            if (country) {
-                searchParams = Object.assign(searchParams, {
-                    componentRestrictions: {
-                        country,
-                    },
-                });
-            }
-
-            searchParams = Object.assign(searchParams, otherSearchOptions);
-        }
         this.setState({ isSearching: true });
 
         return new Promise((resolve, reject) => {
@@ -153,6 +124,7 @@ class PlacesLookupComponent extends Component {
             style,
             label,
             error,
+            readOnly,
             placeholder,
             disabled,
             tabIndex,
@@ -178,7 +150,7 @@ class PlacesLookupComponent extends Component {
                     placeholder={placeholder}
                     options={options}
                     hideLabel={hideLabel}
-                    readOnly={false}
+                    readOnly={readOnly}
                     required={required}
                     tabIndex={tabIndex}
                     disabled={disabled}
@@ -191,11 +163,11 @@ class PlacesLookupComponent extends Component {
                     error={error}
                     icon={<LocationIcon />}
                 />
-                {!error && (
-                    <div className="rainbow-google-address-lookup_poweredby-container">
-                        <PoweredByGoogleLogo className="rainbow-google-address-lookup_poweredby-logo" />
+                <RenderIf isTrue={!error}>
+                    <div className="rainbow-google-address-lookup_powered-by-google-container">
+                        <PoweredByGoogleLogo className="rainbow-google-address-lookup_powered-by-google-logo" />
                     </div>
-                )}
+                </RenderIf>
             </div>
         );
     }
@@ -204,61 +176,11 @@ class PlacesLookupComponent extends Component {
 PlacesLookupComponent.propTypes = {
     isScriptLoaded: PropTypes.bool.isRequired,
     isScriptLoadSucceed: PropTypes.bool.isRequired,
-    searchOptions: PropTypes.shape({
-        bounds: PropTypes.shape({
-            sw: PropTypes.shape({
-                latitude: PropTypes.number.isRequired,
-                longitude: PropTypes.number.isRequired,
-            }),
-            ne: PropTypes.shape({
-                latitude: PropTypes.number.isRequired,
-                longitude: PropTypes.number.isRequired,
-            }),
-        }),
-        country: PropTypes.string,
-        location: PropTypes.shape({
-            latitude: PropTypes.number.isRequired,
-            longitude: PropTypes.number.isRequired,
-        }),
-        radius: PropTypes.number,
-        types: PropTypes.arrayOf(PropTypes.string),
-    }),
+    searchOptions: CustomPropTypes.searchOptionsShape,
     label: PropTypes.oneOfType([PropTypes.string, PropTypes.node]).isRequired,
     hideLabel: PropTypes.bool,
-    value: PropTypes.oneOfType([
-        PropTypes.shape({
-            description: PropTypes.string,
-            id: PropTypes.string,
-            matched_substrings: PropTypes.arrayOf(
-                PropTypes.shape({
-                    length: PropTypes.number,
-                    offset: PropTypes.number,
-                }),
-            ),
-            place_id: PropTypes.string,
-            reference: PropTypes.string,
-            structured_formatting: PropTypes.arrayOf(
-                PropTypes.shape({
-                    main_text: PropTypes.string,
-                    main_text_matched_substrings: PropTypes.arrayOf(
-                        PropTypes.shape({
-                            length: PropTypes.number,
-                            offset: PropTypes.number,
-                        }),
-                    ),
-                    secondary_text: PropTypes.string,
-                }),
-            ),
-            terms: PropTypes.arrayOf(
-                PropTypes.shape({
-                    offset: PropTypes.number,
-                    value: PropTypes.string,
-                }),
-            ),
-            types: PropTypes.arrayOf(PropTypes.string),
-        }),
-        PropTypes.string,
-    ]),
+    readOnly: PropTypes.bool,
+    value: PropTypes.oneOfType([CustomPropTypes.valueShape, PropTypes.string]),
     name: PropTypes.string,
     placeholder: PropTypes.string,
     required: PropTypes.bool,
@@ -280,6 +202,7 @@ PlacesLookupComponent.defaultProps = {
     name: undefined,
     placeholder: null,
     required: false,
+    readOnly: false,
     error: null,
     disabled: false,
     onChange: () => {},
