+++ src/components/GoogleAddressLookup/index.js
@@ -2,6 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import scriptLoader from 'react-async-script-loader';
 import PlacesLookupComponent from './component';
+import * as CustomPropTypes from './proptypes';
+
 /**
  * The GoogleAddressLookup component is used to find a location.
  */
@@ -25,70 +27,7 @@ GoogleAddressLookup.propTypes = {
     /** A boolean to hide the component label. */
     hideLabel: PropTypes.bool,
     /** Specifies the selected value of the component. */
-    value: PropTypes.oneOfType([
-        PropTypes.shape({
-            /**
-             * This is the unformatted version of the query suggested by Google Places service.
-             */
-            description: PropTypes.string,
-            /**
-             * A stable ID for this place, intended to be interoperable with those returned by
-             * the place search service. Note: This has been deprecated in favor of place_id.
-             */
-            id: PropTypes.string,
-            /**
-             * A set of substrings in the place's description that match elements in the user's
-             * input, suitable for use in highlighting those substrings. Each substring is
-             * identified by an offset and a length, expressed in unicode characters.
-             */
-            matched_substrings: PropTypes.arrayOf(
-                PropTypes.shape({
-                    length: PropTypes.number,
-                    offset: PropTypes.number,
-                }),
-            ),
-            /**
-             * A place ID that can be used to retrieve details about this place using the place
-             * details service (see PlacesService.getDetails()).
-             */
-            place_id: PropTypes.string,
-            /**
-             * A reference that can be used to retrieve details about this place using the place
-             * details service (see PlacesService.getDetails()). Note: This has been deprecated
-             * in favor of place_id.
-             */
-            reference: PropTypes.string,
-            structured_formatting: PropTypes.arrayOf(
-                PropTypes.shape({
-                    main_text: PropTypes.string,
-                    main_text_matched_substrings: PropTypes.arrayOf(
-                        PropTypes.shape({
-                            length: PropTypes.number,
-                            offset: PropTypes.number,
-                        }),
-                    ),
-                    secondary_text: PropTypes.string,
-                }),
-            ),
-            /**
-             * Information about individual terms in the above description,
-             * from most to least specific.
-             * For example, "Taco Bell", "Willitis", and "CA".
-             */
-            terms: PropTypes.arrayOf(
-                PropTypes.shape({
-                    offset: PropTypes.number,
-                    value: PropTypes.string,
-                }),
-            ),
-            /**
-             * An array of types that the prediction belongs to,
-             * for example 'establishment' or 'geocode'.
-             */
-            types: PropTypes.arrayOf(PropTypes.string),
-        }),
-        PropTypes.string,
-    ]),
+    value: PropTypes.oneOfType([CustomPropTypes.valueShape, PropTypes.string]),
     /** The name of the component. */
     name: PropTypes.string,
     /** Text that is displayed when the field is empty, to prompt the user for a valid entry. */
@@ -96,6 +35,8 @@ GoogleAddressLookup.propTypes = {
     /** Specifies that the component must be filled out before submitting the form.
      * This value defaults to false. */
     required: PropTypes.bool,
+    /** Specifies that the component is read-only. This value defaults to false. */
+    readOnly: PropTypes.bool,
     /** Specifies that the component must be filled out before submitting the form. */
     error: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
     /** Specifies that the component element should be disabled. This value defaults to false. */
@@ -116,51 +57,20 @@ GoogleAddressLookup.propTypes = {
     onFocus: PropTypes.func,
     /** The action triggered when the element releases focus. */
     onBlur: PropTypes.func,
-    /** */
-    searchOptions: PropTypes.shape({
-        /**
-         * Bounds for prediction biasing. Predictions will be
-         * biased towards, but not restricted to, the given bounds.
-         * Both location and radius will be ignored if bounds is set.
-         */
-        bounds: PropTypes.shape({
-            sw: PropTypes.shape({
-                latitude: PropTypes.number.isRequired,
-                longitude: PropTypes.number.isRequired,
-            }),
-            ne: PropTypes.shape({
-                latitude: PropTypes.number.isRequired,
-                longitude: PropTypes.number.isRequired,
-            }),
-        }),
-        /**
-         * The country in ISO 3166-1 Alpha-2 country code (case insensitive).
-         * E.g. 'us', 'br'
-         */
-        country: PropTypes.string,
-        /**
-         * Location for prediction biasing. Predictions will be biased
-         * towards the given location and radius. Alternatively, bounds
-         * can be used.
-         */
-        location: PropTypes.shape({
-            latitude: PropTypes.number.isRequired,
-            longitude: PropTypes.number.isRequired,
-        }),
-        /**
-         * The radius of the area used for prediction biasing. The radius is
-         * specified in meters, and must always be accompanied by a location
-         * property. Alternatively, bounds can be used.
-         */
-        radius: PropTypes.number,
-        /**
-         * The types of predictions to be returned. Four types are supported:
-         * 'establishment' for businesses, 'geocode' for addresses, '(regions)'
-         * for administrative regions and '(cities)' for localities. If nothing
-         * is specified, all types are returned.
-         */
-        types: PropTypes.arrayOf(PropTypes.string),
-    }),
+    /**
+     * An object with custom search options,
+     * for example:
+     *  searchOptions: {
+     *      location: {
+     *          latitude: -33.941264,
+     *          longitude: 151.2042969,
+     *      },
+     *      country: 'us',
+     *      radius: 150000,
+     *      types: ['address'],
+     *  }
+     */
+    searchOptions: CustomPropTypes.searchOptionsShape,
 };
 
 GoogleAddressLookup.defaultProps = {
