+++ src/components/GoogleAddressLookup/styles.css
@@ -4,13 +4,13 @@
   height: 24px !important;
   font-size: 24px !important; }
 
-.rainbow-google-address-lookup_container .rainbow-google-address-lookup_poweredby-container {
+.rainbow-google-address-lookup_container .rainbow-google-address-lookup_powered-by-google-container {
   display: flex;
   flex-wrap: wrap;
   align-content: flex-start;
   padding-top: 0.25rem;
   padding-right: 0.5rem; }
-  .rainbow-google-address-lookup_container .rainbow-google-address-lookup_poweredby-container .rainbow-google-address-lookup_poweredby-logo {
+  .rainbow-google-address-lookup_container .rainbow-google-address-lookup_powered-by-google-container .rainbow-google-address-lookup_powered-by-google-logo {
     align-self: flex-end;
     margin-left: auto;
     height: 0.625rem; }
