+++ src/components/GoogleAddressLookup/styles.scss
@@ -11,14 +11,14 @@
         }
     }
 
-    .rainbow-google-address-lookup_poweredby-container {
+    .rainbow-google-address-lookup_powered-by-google-container {
         display: flex;
         flex-wrap: wrap;
         align-content: flex-start;
         padding-top: $padding-xx-small;
         padding-right: $padding-x-small;
 
-        .rainbow-google-address-lookup_poweredby-logo {
+        .rainbow-google-address-lookup_powered-by-google-logo {
             align-self: flex-end;
             margin-left: auto;
             height: 0.625rem;
